agent/requirements.txt
# Core dependencies for TaskMaster Agent
langchain
chromadb
llama-cpp-python
pyautogen
requests
pandas
python-dotenv
# For SQLite
sqlite3; python_version >= '3.0'
